+++ src/modules/notes/module.js
@@ -182,7 +182,7 @@ const notesModule = {
     },
   },
   category: "core-workflow",
-  version: "0.33.6.12e-2",
+  version: "0.33.6.12f",
   enabledByDefault: true,
   canDisable: true,
   historicalReadAccess: true,